#180 - Extract bookmark manager to its own class.
Being good!
